Refactor to remove separate package
