+++ README.md
@@ -164,8 +164,7 @@ generated by running `go list -m all`);
 ### resolve-deps
 The `resolve-deps` target compares and merge the dependencies of your plugin module with the dependencies of the Gloo Enterprise one. 
 It will succeed if the shared dependencies match _exactly_ (this is another constraint imposed by Go plugins, more info 
-[here](https://docs.solo.io/gloo/latest/guides/dev/writing_auth_plugins/#build-helper-tools)) and fail otherwise, outputting information 
-about mismatches to stdout.
+[here](https://docs.solo.io/gloo/latest/guides/dev/writing_auth_plugins/#build-helper-tools)) and fail otherwise.
 
 ### build-plugin
 The `build-plugin` target uses the information published by Gloo Enterprise to mirror its build
