+++ pkg/checks/merge_test.go
@@ -76,10 +76,10 @@ var _ = Describe("ParseDependencies", func() {
 					ReplacementVersion: "",
 				},
 				"github.com/solo-io/foo": {
-					Name: "github.com/solo-io/foo",
-					Version: "v0.0.0-20180207000608-0eeff89b0690",
-					Replacement: false,
-					ReplacementName: "",
+					Name:               "github.com/solo-io/foo",
+					Version:            "v0.0.0-20180207000608-0eeff89b0690",
+					Replacement:        false,
+					ReplacementName:    "",
 					ReplacementVersion: "",
 				},
 			},
@@ -94,10 +94,10 @@ var _ = Describe("ParseDependencies", func() {
 					ReplacementVersion: "v1.2.4",
 				},
 				"github.com/solo-io/foo": {
-					Name: "github.com/solo-io/foo",
-					Version: "v0.0.0-20180207000608-0eeff89b0690",
-					Replacement: false,
-					ReplacementName: "",
+					Name:               "github.com/solo-io/foo",
+					Version:            "v0.0.0-20180207000608-0eeff89b0690",
+					Replacement:        false,
+					ReplacementName:    "",
 					ReplacementVersion: "",
 				},
 			},
